Isolate libfmasa.so from libmasa.so
Per bug report from Andy Bauer and subsequent discussion on the
development list, it appears the Fortran libfmasa.so is tangled with
the C/C++ libmasa.so.  Amends...

  * src/Makefile.am so that libfmasa is it's own beast
    ('twould be cleaner to use a convenience library rather
     than to abuse libmasa_la_SOURCES as is done, but this works)

  * Adjusts tests/Makefile.am so that libtool outputs are linked
    rather than using -lmasa and -lfmasa flags directly.  This inherently
    picks up any dependencies from src/ without further work.

I do see

    MASA ERROR!!!:: No such variable  (p_1) exists

    MASA ERROR!!!:: No such variable  (rho_1) exists

    MASA ERROR!!!:: No such variable  (u_1) exists

    MASA ERROR!!!:: No such variable  (w_1) exists

    MASA REGRESSION TEST FAILED!
    Exceeded Threshold by: 1.00094
    FAIL: axi_euler

so I am uncertain that this is sane for merge into master.  In
particular, I do not know that 'make distcheck' will pass since
it is blocked by whatever the issue is with tests/axi_euler on
my system.