AC_REQUIRE compilers, move LT_INIT to the end
This fixes our warnings from Automake.

The constructed configure still tests for Fortran even when we don't
request Fortran, but the documention for AC_REQUIRE swears that
behavior is a feature not a bug: the expanded dependency macro gets
called *before* the dependant macro, even if it was only listed as a
dependency conditionally.

I'm not sure who authored that "feature" so I can't recommend to them
the psychiatric help they need, but it's standard behavior so it's not
easy to escape either.